requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
arm_pytorch_utilities==0.4.3
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cvxpy==1.5.2
cvxpylayers==0.1.6
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
h5py==3.11.0
hydra-core==1.3.2
idna==3.8
importlib_resources==6.4.4
Jinja2==3.1.4
kiwisolver==1.4.5
lightning-utilities==0.11.6
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.161
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
nvitop==1.3.2
omegaconf==2.3.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
protobuf==5.28.0
psutil==6.0.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-seed==0.2.0
pytorch_kinematics==0.7.4
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sympy==1.13.2
termcolor==2.4.0
torch==2.4.1
torchmetrics==1.4.1
torchvision==0.19.1
tqdm==4.66.5
trimesh==4.4.8
triton==3.0.0
typing_extensions==4.12.2
urllib3==2.2.2
viser==0.2.1
wandb==0.17.8
yarl==1.9.6
zipp==3.20.1
